img2vec_service/requirements.txt
fastapi[standard]
img2vec-pytorch-2==1.3.0
python-dotenv